fix: profile layout render
