+++ apps/masterbots.ai/lib/hooks/use-browse.tsx
@@ -1,3 +1,4 @@
+// ! TODO: Remove this hook and related components, not in use. @andlerrl 👀
 'use client'
 
 import * as React from 'react'
